+++ src/main/java/refinedstorage/inventory/IItemValidator.java
@@ -3,5 +3,5 @@ package refinedstorage.inventory;
 import net.minecraft.item.ItemStack;
 
 public interface IItemValidator {
-    boolean valid(ItemStack stack);
+    boolean isValid(ItemStack stack);
 }