+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fixed External Storage disconnecting on world reload
 - Fixed External Storage not updating correctly
 - Fixed wireless signal starting from Controller instead of per Wireless Transmitter individually
+- Fixed Controller's redstone state not saving
 - Huge performance improvements to large storage networks
 
 **Features**